binary: Alter pico-8 font height.
I transformed the glyphs in this font using FontForge by scaling them
all by 1.2 on x and y. This removed the extra "pixel" at the top of the
font that was causing Bevy's renderer to add too much height to each
line when used with newlines. This does mean the font is larger than it
was. I don't know enough about fonts to understand whether its size
ought to adhere to some standard. I decremented the font size by one in
the code and it looks like what I want.